Squash a sign conversion warning

+++ cb.c
@@ -124,7 +124,7 @@ Tsetl(Tbl *tbl, const char *key, size_t len, void *val) {
 	return(tbl);
 newkey:; // We have the byte index; what about the bit?
 	uint k1 = (byte)key[i], k2 = (byte)t->leaf.key[i];
-	uint b = __builtin_clz((k1 ^ k2) << 24 | 0x800000);
+	uint b = (uint)__builtin_clz((k1 ^ k2) << 24 | 0x800000);
 	i = 8 * i + b;
 	b = k1 >> (7 - b) & 1;
 	// Find where to insert a branch or grow an existing branch.
